# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Guest Service API",
        "description": "API untuk manajemen data tamu hotel. NIM: 102022400289",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/guests": {
            "get": {
                "tags": [
                    "Guest Service (Data Diri Tamu)"
                ],
                "summary": "Ambil daftar semua guest",
                "description": "Collection: GET /api/v1/guests\nMengambil daftar semua guest.",
                "operationId": "2a4071b3b7b64836d4f7ff4cf5bdbe67",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Guest Service (Data Diri Tamu)"
                ],
                "summary": "Tambah data guest baru",
                "description": "Action: POST /api/v1/guests\nMenambah data guest baru.",
                "operationId": "ba7114d5bd8184b0149e88e634336bea",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Aurel"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "aurel@example.com"
                                    },
                                    "ktp_number": {
                                        "type": "string",
                                        "example": "1234567890123456"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "08123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/guests/{id}": {
            "get": {
                "tags": [
                    "Guest Service (Data Diri Tamu)"
                ],
                "summary": "Ambil profil guest berdasarkan ID",
                "description": "Resource: GET /api/v1/guests/{id}\nMengambil data guest spesifik berdasarkan ID.",
                "operationId": "cb620592794746fd435ead18f15ffc4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Guest Service (Data Diri Tamu)",
            "description": "Guest Service (Data Diri Tamu)"
        }
    ]
}